fix: move shared code from file_unix.go
