+++ pkg/file/file_unix.go
@@ -4,12 +4,8 @@ package file
 
 import (
 	"errors"
-	"fmt"
-	"io"
 	"os"
 	"syscall"
-
-	errors2 "github.com/influxdata/influxdb/pkg/errors"
 )
 
 func SyncDir(dirName string) error {
@@ -57,39 +53,4 @@ func RenameFileWithReplacement(oldpath, newpath string) error {
 // be identical to oldpath, and oldpath will be removed.
 func RenameFile(oldpath, newpath string) error {
 	return RenameFileWithReplacement(oldpath, newpath)
-}
-
-func copyFile(src, dst string) (err error) {
-	in, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-
-	out, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-
-	defer errors2.Capture(&err, out.Close)()
-
-	defer errors2.Capture(&err, in.Close)()
-
-	if _, err = io.Copy(out, in); err != nil {
-		return err
-	}
-
-	return out.Sync()
-}
-
-// MoveFileWithReplacement copies the file contents at `src` to `dst`.
-// and deletes `src` on success.
-//
-// If the file at `dst` already exists, it will be truncated and its contents
-// overwritten.
-func MoveFileWithReplacement(src, dst string) error {
-	if err := copyFile(src, dst); err != nil {
-		return fmt.Errorf("copy: %w", err)
-	}
-
-	return os.Remove(src)
 }
